REST api suggestion (#1)
Reuse the same url for doctor GET/PUT/DELETE. The url doesn't have to be unique if the method is different. You can follow this pattern for patients and appointments as well.
Switch admin and public urls. This makes it easier to share the url with the public and harder to guess the admin url which we want to limit access to.